+++ src/Ext/Techno/Body.cpp
@@ -498,6 +498,7 @@ void TechnoExt::ExtData::Serialize(T& Stm)
 		.Process(this->AE_Cloakable)
 		.Process(this->AE_ForceDecloak)
 		.Process(this->AE_DisableWeapons)
+		.Process(this->LastWeaponFLH)
 		.Process(this->FiringObstacleCell)
 		;
 }